Move username to admin panel
