+++ apps/obsidian/src/components/AdminPanelSettings.tsx
@@ -1,12 +1,16 @@
 import { useState, useCallback } from "react";
 import { usePlugin } from "./PluginContext";
 import { Notice } from "obsidian";
+import { updateUsername } from "~/utils/supabaseContext";
 import { initializeSupabaseSync } from "~/utils/syncDgNodesToSupabase";
 
 export const AdminPanelSettings = () => {
   const plugin = usePlugin();
   const [syncModeEnabled, setSyncModeEnabled] = useState<boolean>(
     plugin.settings.syncModeEnabled ?? false,
+  );
+  const [username, setUsername] = useState<string>(
+    plugin.settings.username || "",
   );
 
   const handleSyncModeToggle = useCallback(
@@ -30,6 +34,13 @@ export const AdminPanelSettings = () => {
     [plugin],
   );
 
+  const handleSetUsername = async (newValue: string) => {
+    setUsername(newValue);
+    plugin.settings.username = newValue;
+    await plugin.saveSettings();
+    await updateUsername(plugin, newValue);
+  };
+
   return (
     <div className="general-settings">
       <div className="setting-item">
@@ -46,6 +57,27 @@ export const AdminPanelSettings = () => {
           >
             <input type="checkbox" checked={syncModeEnabled} />
           </div>
+        </div>
+      </div>
+      <div
+        className={
+          "setting-item " + (plugin.settings.syncModeEnabled ? "" : "hidden")
+        }
+      >
+        <div className="setting-item-info">
+          <div className="setting-item-name">Username</div>
+          <div className="setting-item-description">
+            A username that will be associated with your vault if you share
+            data.
+          </div>
+        </div>
+        <div className="setting-item-control">
+          <input
+            type="text"
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+            onBlur={(e) => void handleSetUsername(e.target.value)}
+          />
         </div>
       </div>
     </div>
